Adapt the height of the block preview container  to its content
