+++ packages/block-editor/src/components/block-preview/style.scss
@@ -29,18 +29,17 @@
 	transform-origin: top left;
 
 	// Resetting paddings, margins, and other.
+
 	text-align: initial;
 	margin: 0;
 	overflow: visible;
 	min-height: auto;
 
-	.block-editor-block-list__layout,
-	.block-editor-block-list__block {
-		padding: 0;
-	}
-
-	> div section {
-		height: auto;
+	.block-editor-block-preview__container &.is-centered {
+		.block-editor-block-list__layout,
+		.block-editor-block-list__block {
+			padding: 0;
+		}
 	}
 
 	.block-editor-block-list__insertion-point,
